+++ examples/routes/foo-[id].js
@@ -12,9 +12,7 @@ export default class Params extends LitElement {
     return html`
       <h1>Params</h1>
       <p>SSR content with params: ${JSON.stringify(this.ctx)}</p>
-      <is-land
-        ><island-bar .ctx=${this.ctx} .foo=${this.foo}></island-bar
-      ></is-land>
+      <island-bar .ctx=${this.ctx} .foo=${this.foo}></island-bar>
       <a href="/">To home</a>
     `;
   }