+++ examples/routes/index.js
@@ -7,7 +7,7 @@ export default class Home extends LitElement {
       <section class="p-6">
         <h1>Home</h1>
         <p class="text-2xl font-bold text-blue-600/100">SSR content</p>
-        <is-land><island-foo name="Iris"></island-foo></is-land>
+        <island-foo name="Iris"></island-foo>
 
         <section>
           <a href="/foo/bar">To foo/bar</a> | <a href="/foo-123">To params</a> |