+++ src/mod.ts
@@ -3,9 +3,14 @@ export { Builder } from "./dev/builder.ts";
 export { fsRoutes } from "./plugins/fs-routes.ts";
 export { tailwind } from "./plugins/tailwind.ts";
 export { staticFiles } from "./server/static-files.ts";
+export { ContextMixin } from "./server/context.ts";
 export type { Context } from "./server/context.ts";
 export type { AppWrapperComponent, AppWrapperOptions } from "./server/ssr.ts";
-export type { RouteConfig } from "./server/router.ts";
+export type { RouteConfig, RouteModule } from "./server/router.ts";
 export type { Handlers } from "./server/handlers.ts";
 export type { MiddlewareFn } from "./server/middlewares.ts";
-export type { LayoutConfig, LayoutComponent } from "./server/layouts.ts";
+export type {
+  LayoutConfig,
+  LayoutComponent,
+  LayoutModule,
+} from "./server/layouts.ts";